Refactored type resolver to reflection namespace
