+++ Eventum.sln
@@ -52,6 +52,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eventum.Telemetry.OpenTelem
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eventum.Telemetry.OpenTelemetry.Tests", "tests\Eventum.Telemetry.OpenTelemetry.Tests\Eventum.Telemetry.OpenTelemetry.Tests.csproj", "{D7241207-2C09-4750-8673-FA60E3A18186}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eventum.Reflection", "src\Eventum.Reflection\Eventum.Reflection.csproj", "{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eventum.Reflection.Tests", "tests\Eventum.Reflection.Tests\Eventum.Reflection.Tests.csproj", "{43FCA04C-FB70-462C-89D0-99C09A2DA63D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{D85CE62A-BEBD-46BB-A0B0-B4EEEAEC2179}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -122,6 +128,14 @@ Global
 		{D7241207-2C09-4750-8673-FA60E3A18186}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D7241207-2C09-4750-8673-FA60E3A18186}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D7241207-2C09-4750-8673-FA60E3A18186}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{43FCA04C-FB70-462C-89D0-99C09A2DA63D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{43FCA04C-FB70-462C-89D0-99C09A2DA63D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{43FCA04C-FB70-462C-89D0-99C09A2DA63D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{43FCA04C-FB70-462C-89D0-99C09A2DA63D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -143,6 +157,8 @@ Global
 		{4ADBCF13-CFD2-44F8-8410-32C6C83AA9E7} = {BD127C35-01F5-4C15-A295-36EC25EAA113}
 		{E0573F56-6A04-47C0-814E-B25B1C74FAB5} = {BD127C35-01F5-4C15-A295-36EC25EAA113}
 		{D7241207-2C09-4750-8673-FA60E3A18186} = {BD127C35-01F5-4C15-A295-36EC25EAA113}
+		{9A2A92BA-248E-45D7-BD6C-ACAC9322DFDC} = {D85CE62A-BEBD-46BB-A0B0-B4EEEAEC2179}
+		{43FCA04C-FB70-462C-89D0-99C09A2DA63D} = {D85CE62A-BEBD-46BB-A0B0-B4EEEAEC2179}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4AFD14DC-99D7-40B7-B63F-F48B97B1B86E}
