+++ src/Eventum.Persistence.CosmosDb/Eventum.Persistence.CosmosDb.csproj
@@ -13,6 +13,7 @@
     <ProjectReference Include="..\Eventum.EventSourcing.Abstractions\Eventum.EventSourcing.Abstractions.csproj" />
     <ProjectReference Include="..\Eventum.Persistence.Abstractions\Eventum.Persistence.Abstractions.csproj" />
     <ProjectReference Include="..\Eventum.Projection.Abstractions\Eventum.Projection.Abstractions.csproj" />
+    <ProjectReference Include="..\Eventum.Reflection\Eventum.Reflection.csproj" />
   </ItemGroup>
 
 </Project>
