+++ tests/Eventum.Reflection.Tests/Eventum.Reflection.Tests.csproj
@@ -2,11 +2,16 @@
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Eventum.Reflection\Eventum.Reflection.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.46.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
     <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="xunit" Version="2.9.3" />
@@ -21,8 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Eventum.EventSourcing.Abstractions\Eventum.EventSourcing.Abstractions.csproj" />
-    <ProjectReference Include="..\..\src\Eventum.Reflection\Eventum.Reflection.csproj" />
+    <Folder Include="TestData\" />
   </ItemGroup>
 
 </Project>
